requirements.txt
django==4.0
django-rest-framework==3.14
moviepy==1.0